feat(routes): change extesion
