+++ frontend/src/help/request.tsx
@@ -1,0 +1,8 @@
+export const getUserInfo = async (url: string): Promise<string> => {
+  return (
+    await fetch(url)
+      .then(async (response) => await response.json())
+      // and return the result data.
+      .then((data) => data)
+  )
+}
